Inline Case files for UtilityClassWithPublicConstructorSpec
